Do not set (SO)VERSION on Windows DLLs.
The SYCL DLL building code does not work correctly with the
latest CMake versions when versioning information is set. :-(

There have been issues with this since forever...

https://discourse.cmake.org/t/custom-language-setup-on-windows/4194

It's just that with the latest CMake versions the dependent code
does not link correctly against vecmem_sycl.dll anymore when
versioning information is set on it.