Introduced vecmem::edm::container.
It is a helper class for declaring all related types for a given
EDM schema, a bit more conveniently.

Re-wrote the tests to use this new formalism, on a slightly more
complicated container than was used before. As a result, fixed
some minor bugs in the code that showed up with the more complicated
EDM.